microservice_scout/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
firebase-admin>=6.5.0
requests>=2.31.0
python-dotenv>=1.0.0
# --- FASE 2: Radar Intel ---
pytrends>=4.9.2
feedparser>=6.0.10
beautifulsoup4>=4.12.3
lxml>=5.1.0
pandas>=2.2.0
urllib3==1.26.15
# --- FASE 3: Ojos de Depredador ---
httpx>=0.27.0
opencv-python-headless>=4.9.0.80
pytesseract>=0.3.10
fer>=22.5.1
tensorflow-cpu>=2.15.0  # Versión CPU para mantener la imagen ligera
fake-useragent>=1.4.0
aiofiles>=23.2.1